beartell/requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv
minio
python-magic
python-magic-bin
pgvector
openai
tiktoken
pypdf2
python-docx
pdfplumber
pytesseract
Pillow
